requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
